app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "savaobay", email = "savaobay2@gmail.com" }]
dependencies = [
	"pyserial>=3.5",
	"numpy>=1.26.4",
	"pillow>=10.3.0",
	"ruff>=0.6.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
builtins = ["_"]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
